+++ proto/CHANGELOG.md
@@ -3,18 +3,23 @@
 The original files come from https://github.com/anod/headunit; 
 this file documents the changes made to the protobuf files.
 
-# 2026-01-31
+# 2026-02-01 (F-Jan)
+- Added more VideoResolutions
+- Added the DisplayType enum
+- Made `metastate` of the `Key` message required
+
+# 2026-01-31 (F-Jan)
 - Changed field name `driver_position` to `left_hand_traffic`
 - Added new MessageStatus types
 
-# 2026-01-25
+# 2026-01-25 (F-Jan)
 - fixed some typos
   - `color_depth_bits`
   - `album_art`
   - `parking_brake_data`
   - `environment_data`
 
-# 2026-01-24
+# 2026-01-24 (F-Jan)
 - Changed Enum naming to Rust naming
 - Fixed some typos
 - Removed java specific parts